moved CreateWalletPin to src/pages
